requirements.txt
# Core dependencies
scapy>=2.4.5
flask>=2.0.0
netifaces>=0.11.0
pycryptodome>=3.10.0
qrcode>=7.0.0
pillow>=8.3.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pyyaml>=5.4.0
colorama>=0.4.4
tabulate>=0.8.0
cryptography>=3.4.0

# For Windows development (will be ignored on Linux)
pybluez>=0.23; sys_platform == 'win32'

# Optional advanced features
selenium>=4.0.0
pymongo>=3.11.0